requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
pytz==2023.3
gunicorn==21.2.0